fix comments & move preprocessing under layers
