+++ keras_nlp/layers/preprocessing/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_nlp.preprocessing.mlm_mask_generator import MLMMaskGenerator
+from keras_nlp.layers.preprocessing.mlm_mask_generator import MLMMaskGenerator
