+++ keras_nlp/layers/preprocessing/mlm_mask_generator_test.py
@@ -14,7 +14,7 @@
 
 import tensorflow as tf
 
-from keras_nlp.preprocessing.mlm_mask_generator import MLMMaskGenerator
+from keras_nlp.layers.preprocessing.mlm_mask_generator import MLMMaskGenerator
 
 
 class MLMMaskGeneratorTest(tf.test.TestCase):
